Fix TypeScript errors and switch to SQLite for local development
- Fix unused variable errors in backend (prefix with _)
- Fix type assertion errors in frontend (bigint handling)
- Fix null safety checks in hooks and components
- Switch from PostgreSQL to SQLite for local development
- Remove duplicate frontend/pnpm-lock.yaml
- Clean up pnpm-workspace.yaml formatting

All builds now pass with no errors or warnings.

+++ frontend/app/my-books/page.tsx
@@ -23,7 +23,7 @@ function BorrowedBookCard({ bookId }: { bookId: bigint }) {
   };
 
   const dueDate = borrowedUntil ? new Date(Number(borrowedUntil) * 1000) : null;
-  const isOverdue = borrowedUntil && borrowedUntil < BigInt(Math.floor(Date.now() / 1000));
+  const isOverdue = borrowedUntil && (borrowedUntil as bigint) < BigInt(Math.floor(Date.now() / 1000));
 
   if (!isBorrowed) return null;
